dom/base/test/test_bug769117.html

<!DOCTYPE HTML>
<html>
  <!--
       https://bugzilla.mozilla.org/show_bug.cgi?id=769117
     -->
  <head>
    <meta charset="utf-8">
    <title>Test for Bug 769117</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    <script type="application/javascript">
     SimpleTest.waitForExplicitFinish();
     /** Test for Bug 769117 **/
     "use strict";
     function onLoad () {
       SpecialPowers.pushPrefEnv({"set": [["plugins.rewrite_youtube_embeds", true]]}, function() {

         let youtube_url = "https://mochitest.youtube.com/v/Xm5i5kbIXzc";
         let youtube_changed_url = "https://mochitest.youtube.com/embed/Xm5i5kbIXzc";
         let static_iframe = document.getElementById("staticiframe");

         function testEmbed(embed) {
           ok (embed, "Embed node exists");
           embed = SpecialPowers.wrap(embed);
           is (embed.srcURI.spec, youtube_changed_url, "Should have src uri of " + youtube_changed_url);
         }

         function testStatic() {
           info("Running static embed youtube rewrite test");
           iframe_doc = static_iframe.contentWindow.document;
           testEmbed(iframe_doc.getElementById("testembed"));
           testEmbed(iframe_doc.getElementById("testobject"));
           SimpleTest.executeSoon(() => {
             testEmbed(embed_dynamic);
             SimpleTest.finish();
           });
         }

         info("Running dynamic embed youtube rewrite test");
         let embed_dynamic = document.createElement("embed");
         embed_dynamic.src = "https://mochitest.youtube.com/v/Xm5i5kbIXzc";
         embed_dynamic.type = "application/x-shockwave-flash";
         document.body.appendChild(embed_dynamic);

         static_iframe.onload = testStatic;
         static_iframe.src = "file_bug769117.html"

       });
     }
    </script>
  </head>
  <body onload="onLoad()">
    <iframe id="staticiframe"></iframe>
  </body>
</html>
